+++ components/Chat/Thinking.tsx
@@ -1,22 +1,16 @@
 import { LoaderCircle, TvMinimalPlay } from "lucide-react";
+import useProfileSerach from "@/hooks/useProfileSearch";
 import getZoraPfpLink from "@/lib/zora/getZoraPfpLink";
-import { useProfileProvider } from "@/providers/ProfileProvider";
 
 const Thinking = () => {
-  const { profile } = useProfileProvider();
+  const { profile } = useProfileSerach();
 
   return (
     <div className="flex gap-2 w-full max-w-3xl mx-auto items-center pb-2 text-black">
       <div className="size-fit">
         {
-          profile ? (
-            <img
-              src={getZoraPfpLink(profile)}
-              alt="PFP"
-              width={36}
-              height={36}
-              className="rounded-full"
-            />
+          profile.length > 0 ? (
+            <img src={getZoraPfpLink(profile[0])} alt="PFP" width={36} height={36} className="rounded-full" />
           ) : (
             <TvMinimalPlay className="h-6 w-6" />
           )
